Split out lib.ts
